requirements.txt
selenium
webdriver-manager
python-docx
